Fix multi-threaded pool with a semaphore
